+++ src/main/java/domain/board/Position.java
@@ -1,3 +1,5 @@
+package domain.board;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
