+++ src/main/java/domain/board/Row.java
@@ -1,3 +1,5 @@
+package domain.board;
+
 import java.util.Arrays;
 
 public enum Row {
