+++ src/main/java/domain/piece/EmptyPiece.java
@@ -1,4 +1,7 @@
-import java.awt.*;
+package domain.piece;
+
+import domain.board.Position;
+
 import java.util.List;
 
 public class EmptyPiece extends Piece {
