+++ src/main/java/domain/piece/Guard.java
@@ -1,31 +1,29 @@
+package domain.piece;
+
+import domain.board.Position;
+
 import java.util.List;
 
-public class Soldier extends Piece {
+public class Guard extends Piece {
 
-    public Soldier(PieceColor color) {
+    public Guard(PieceColor color) {
         super(color);
     }
 
     @Override
     public boolean isValidDestination(Position source, Position destination) {
-        int rowDifference = source.rowDifference(destination);
-        int columnDifference = source.columnDifference(destination);
-
-        if ((color == PieceColor.RED) && (rowDifference == 1 && columnDifference == 0)) {
-            return true;
-        }
-        if ((color == PieceColor.BLUE) && (rowDifference == -1 && columnDifference == 0)) {
-            return true;
-        }
-        if ((rowDifference == 0 && columnDifference == -1) || (rowDifference == 0 && columnDifference == 1)) {
-            return true;
-        }
         return false;
     }
 
     @Override
     public List<Position> findAllRoute(Position source, Position destination) {
         return List.of();
+    }
+
+
+    @Override
+    public boolean isOtherTeam(Piece other) {
+        return super.isOtherTeam(other);
     }
 
     @Override
