+++ src/main/java/domain/piece/Piece.java
@@ -1,3 +1,7 @@
+package domain.piece;
+
+import domain.board.Position;
+
 import java.util.List;
 
 public abstract class Piece {
