+++ src/main/java/domain/piece/PieceColor.java
@@ -1,3 +1,5 @@
+package domain.piece;
+
 public enum PieceColor {
     RED,
     BLUE,
