+++ src/test/java/domain/board/PositionTest.java
@@ -1,3 +1,5 @@
+package domain.board;
+
 import org.junit.jupiter.api.Test;
 
 import java.util.List;
